Added fix for double type

+++ lib/data_value.dart
@@ -1,10 +1,10 @@
 class DataValue {
-  int timestamp;
+  double timestamp;
   double value;
 
   DataValue({required this.timestamp, required this.value});
 
   factory DataValue.fromBuffer(Map<String, dynamic> data) {
-    return DataValue(timestamp: data['timestamp'] as int, value: data['value']);
+    return DataValue(timestamp: data['timestamp'], value: data['value']);
   }
 }
